Add font options and refactor BaseClockScreen time display
This commit introduces new font options for the clock display. It also refactors the BaseClockScreen to enhance how the time is displayed, improving readability and user experience.

This also changes how the background animations are built namely the pixel fire animation.

This also adds a new logo for the app.